Java files for tree implementation
These are files from Netbeans. I imagine they're openable from most Java IDEs. The tree still needs some work and I haven't really tested it yet. But the basics are there and it should function. I think the main thing left, is to perhaps implement an auto numbering for nodes in a tree, so that they're easily accessible.